scheduler: extract `feasible`
