+++ scheduler/benchmarks/helpers_test.go
@@ -15,7 +15,7 @@ import (
 	"testing"
 
 	"github.com/hashicorp/nomad/helper/raftutil"
-	"github.com/hashicorp/nomad/scheduler"
+	"github.com/hashicorp/nomad/scheduler/tests"
 	"github.com/shoenig/test/must"
 )
 
@@ -25,7 +25,7 @@ import (
 // - NOMAD_BENCHMARK_DATADIR: path to data directory
 // - NOMAD_BENCHMARK_SNAPSHOT: path to raft snapshot
 // - neither: empty starting state
-func NewBenchmarkingHarness(t testing.TB) *scheduler.Harness {
+func NewBenchmarkingHarness(t testing.TB) *tests.Harness {
 	// create the Harness and starting state.
 	datadir := os.Getenv("NOMAD_BENCHMARK_DATADIR")
 	if datadir != "" {
@@ -40,12 +40,12 @@ func NewBenchmarkingHarness(t testing.TB) *scheduler.Harness {
 			return h
 		}
 	}
-	return scheduler.NewHarness(t)
+	return tests.NewHarness(t)
 }
 
 // NewHarnessFromDataDir creates a new scheduler test harness with
 // state loaded from an existing datadir.
-func NewHarnessFromDataDir(t testing.TB, datadirPath string) (*scheduler.Harness, error) {
+func NewHarnessFromDataDir(t testing.TB, datadirPath string) (*tests.Harness, error) {
 	if datadirPath == "" {
 		return nil, errors.New("datadir path was not set")
 	}
@@ -58,12 +58,12 @@ func NewHarnessFromDataDir(t testing.TB, datadirPath string) (*scheduler.Harness
 		return nil, err
 	}
 
-	return scheduler.NewHarnessWithState(t, fsm.State()), nil
+	return tests.NewHarnessWithState(t, fsm.State()), nil
 }
 
 // NewHarnessFromDataDir creates a new harness with state loaded
 // from an existing raft snapshot.
-func NewHarnessFromSnapshot(t testing.TB, snapshotPath string) (*scheduler.Harness, error) {
+func NewHarnessFromSnapshot(t testing.TB, snapshotPath string) (*tests.Harness, error) {
 	if snapshotPath == "" {
 		return nil, errors.New("snapshot path was not set")
 	}
@@ -78,5 +78,5 @@ func NewHarnessFromSnapshot(t testing.TB, snapshotPath string) (*scheduler.Harne
 		return nil, err
 	}
 
-	return scheduler.NewHarnessWithState(t, state), nil
+	return tests.NewHarnessWithState(t, state), nil
 }
