+++ scheduler/generic_sched_test.go
@@ -19,6 +19,8 @@ import (
 	"github.com/hashicorp/nomad/nomad/mock"
 	"github.com/hashicorp/nomad/nomad/structs"
 	"github.com/hashicorp/nomad/scheduler/reconcile"
+	sstructs "github.com/hashicorp/nomad/scheduler/structs"
+	"github.com/hashicorp/nomad/scheduler/tests"
 	"github.com/shoenig/test"
 	"github.com/shoenig/test/must"
 )
@@ -26,7 +28,7 @@ import (
 func TestServiceSched_JobRegister(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	for range 10 {
@@ -126,7 +128,7 @@ func TestServiceSched_JobRegister(t *testing.T) {
 func TestServiceSched_JobRegister_StickyAllocs(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	for i := 0; i < 10; i++ {
@@ -182,7 +184,7 @@ func TestServiceSched_JobRegister_StickyAllocs(t *testing.T) {
 		Status:      structs.EvalStatusPending,
 	}
 	must.NoError(t, h.State.UpsertEvals(structs.MsgTypeTestSetup, h.NextIndex(), []*structs.Evaluation{eval}))
-	h1 := NewHarnessWithState(t, h.State)
+	h1 := tests.NewHarnessWithState(t, h.State)
 	if err := h1.Process(NewServiceScheduler, eval); err != nil {
 		t.Fatalf("err: %v", err)
 	}
@@ -220,7 +222,7 @@ func TestServiceSched_JobRegister_StickyAllocs(t *testing.T) {
 func TestServiceSched_JobRegister_StickyHostVolumes(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	nodes := []*structs.Node{
 		mock.Node(),
@@ -329,7 +331,7 @@ func TestServiceSched_JobRegister_StickyHostVolumes(t *testing.T) {
 func TestServiceSched_JobRegister_DiskConstraints(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a node
 	node := mock.Node()
@@ -405,7 +407,7 @@ func TestServiceSched_JobRegister_DiskConstraints(t *testing.T) {
 func TestServiceSched_JobRegister_DistinctHosts(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	for i := 0; i < 10; i++ {
@@ -489,7 +491,7 @@ func TestServiceSched_JobRegister_DistinctHosts(t *testing.T) {
 func TestServiceSched_JobRegister_DistinctProperty(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	for i := 0; i < 10; i++ {
@@ -588,7 +590,7 @@ func TestServiceSched_JobRegister_DistinctProperty(t *testing.T) {
 func TestServiceSched_JobRegister_DistinctProperty_TaskGroup(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	for i := 0; i < 2; i++ {
@@ -669,7 +671,7 @@ func TestServiceSched_JobRegister_DistinctProperty_TaskGroup(t *testing.T) {
 func TestServiceSched_JobRegister_DistinctProperty_TaskGroup_Incr(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a job that uses distinct property over the node-id
 	job := mock.Job()
@@ -758,7 +760,7 @@ func TestServiceSched_Spread(t *testing.T) {
 	for i := 0; i < 10; i++ {
 		name := fmt.Sprintf("%d%% in dc1", start)
 		t.Run(name, func(t *testing.T) {
-			h := NewHarness(t)
+			h := tests.NewHarness(t)
 			remaining := uint8(100 - start)
 			// Create a job that uses spread over data center
 			job := mock.Job()
@@ -853,7 +855,7 @@ func TestServiceSched_Spread(t *testing.T) {
 func TestServiceSched_JobRegister_Datacenter_Downgrade(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create 5 nodes in each datacenter.
 	// Use two loops so nodes are separated by datacenter.
@@ -963,7 +965,7 @@ func TestServiceSched_JobRegister_Datacenter_Downgrade(t *testing.T) {
 func TestServiceSched_JobRegister_NodePool_Downgrade(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Set global scheduler configuration.
 	h.State.SchedulerSetConfig(h.NextIndex(), &structs.SchedulerConfiguration{
@@ -1094,7 +1096,7 @@ func TestServiceSched_JobRegister_NodePool_Downgrade(t *testing.T) {
 func TestServiceSched_EvenSpread(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 	// Create a job that uses even spread over data center
 	job := mock.Job()
 	job.Datacenters = []string{"dc1", "dc2"}
@@ -1167,7 +1169,7 @@ func TestServiceSched_EvenSpread(t *testing.T) {
 func TestServiceSched_JobRegister_Annotate(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	for i := 0; i < 10; i++ {
@@ -1247,7 +1249,7 @@ func TestServiceSched_JobRegister_Annotate(t *testing.T) {
 func TestServiceSched_JobRegister_CountZero(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	for i := 0; i < 10; i++ {
@@ -1298,7 +1300,7 @@ func TestServiceSched_JobRegister_CountZero(t *testing.T) {
 func TestServiceSched_JobRegister_AllocFail(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create NO nodes
 	// Create a job
@@ -1376,7 +1378,7 @@ func TestServiceSched_JobRegister_AllocFail(t *testing.T) {
 func TestServiceSched_JobRegister_CreateBlockedEval(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a full node
 	node := mock.Node()
@@ -1478,7 +1480,7 @@ func TestServiceSched_JobRegister_CreateBlockedEval(t *testing.T) {
 func TestServiceSched_JobRegister_FeasibleAndInfeasibleTG(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create one node
 	node := mock.Node()
@@ -1626,7 +1628,7 @@ func TestServiceSched_JobRegister_SchedulerAlgorithm(t *testing.T) {
 	for _, jobType := range jobTypes {
 		for _, tc := range testCases {
 			t.Run(fmt.Sprintf("%s/%s", jobType, tc.name), func(t *testing.T) {
-				h := NewHarness(t)
+				h := tests.NewHarness(t)
 
 				// Create node pools.
 				nodePools := []*structs.NodePool{
@@ -1678,14 +1680,14 @@ func TestServiceSched_JobRegister_SchedulerAlgorithm(t *testing.T) {
 				}
 				must.NoError(t, h.State.UpsertEvals(structs.MsgTypeTestSetup, h.NextIndex(), []*structs.Evaluation{eval}))
 
-				var scheduler Factory
+				var sched sstructs.Factory
 				switch jobType {
 				case "batch":
-					scheduler = NewBatchScheduler
+					sched = NewBatchScheduler
 				case "service":
-					scheduler = NewServiceScheduler
+					sched = NewServiceScheduler
 				}
-				err := h.Process(scheduler, eval)
+				err := h.Process(sched, eval)
 				must.NoError(t, err)
 
 				must.Len(t, 1, h.Plans)
@@ -1703,7 +1705,7 @@ func TestServiceSched_JobRegister_SchedulerAlgorithm(t *testing.T) {
 					Status:      structs.EvalStatusPending,
 				}
 				must.NoError(t, h.State.UpsertEvals(structs.MsgTypeTestSetup, h.NextIndex(), []*structs.Evaluation{eval}))
-				err = h.Process(scheduler, eval)
+				err = h.Process(sched, eval)
 				must.NoError(t, err)
 
 				must.Len(t, 2, h.Plans)
@@ -1736,7 +1738,7 @@ func TestServiceSched_JobRegister_SchedulerAlgorithm(t *testing.T) {
 func TestServiceSched_EvaluateMaxPlanEval(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a job and set the task group count to zero.
 	job := mock.Job()
@@ -1773,10 +1775,10 @@ func TestServiceSched_EvaluateMaxPlanEval(t *testing.T) {
 func TestServiceSched_Plan_Partial_Progress(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a node of limited resources
-	legacyCpuResources4000, processorResources4000 := cpuResources(4000)
+	legacyCpuResources4000, processorResources4000 := tests.CpuResources(4000)
 	node := mock.Node()
 	node.NodeResources.Processors = processorResources4000
 	node.NodeResources.Cpu = legacyCpuResources4000
@@ -1836,7 +1838,7 @@ func TestServiceSched_Plan_Partial_Progress(t *testing.T) {
 func TestServiceSched_EvaluateBlockedEval(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a job
 	job := mock.Job()
@@ -1883,7 +1885,7 @@ func TestServiceSched_EvaluateBlockedEval(t *testing.T) {
 func TestServiceSched_EvaluateBlockedEval_Finished(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	for i := 0; i < 10; i++ {
@@ -1970,7 +1972,7 @@ func TestServiceSched_EvaluateBlockedEval_Finished(t *testing.T) {
 func TestServiceSched_JobModify(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -2075,7 +2077,7 @@ func TestServiceSched_JobModify(t *testing.T) {
 func TestServiceSched_JobModify_ExistingDuplicateAllocIndex(t *testing.T) {
 	ci.Parallel(t)
 
-	testHarness := NewHarness(t)
+	testHarness := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -2152,7 +2154,7 @@ func TestServiceSched_JobModify_ExistingDuplicateAllocIndex(t *testing.T) {
 func TestServiceSched_JobModify_ProposedDuplicateAllocIndex(t *testing.T) {
 	ci.Parallel(t)
 
-	testHarness := NewHarness(t)
+	testHarness := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -2284,7 +2286,7 @@ func TestServiceSched_JobModify_ProposedDuplicateAllocIndex(t *testing.T) {
 func TestServiceSched_JobModify_ExistingDuplicateAllocIndexNonDestructive(t *testing.T) {
 	ci.Parallel(t)
 
-	testHarness := NewHarness(t)
+	testHarness := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -2374,7 +2376,7 @@ func TestServiceSched_JobModify_ExistingDuplicateAllocIndexNonDestructive(t *tes
 func TestServiceSched_JobModify_Datacenters(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes in 3 DCs
 	var nodes []*structs.Node
@@ -2454,7 +2456,7 @@ func TestServiceSched_JobModify_Datacenters(t *testing.T) {
 func TestServiceSched_JobModify_IncrCount_NodeLimit(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create one node
 	node := mock.Node()
@@ -2548,7 +2550,7 @@ func TestServiceSched_JobModify_IncrCount_NodeLimit(t *testing.T) {
 func TestServiceSched_JobModify_CountZero(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -2650,7 +2652,7 @@ func TestServiceSched_JobModify_CountZero(t *testing.T) {
 func TestServiceSched_JobModify_Rolling(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -2758,7 +2760,7 @@ func TestServiceSched_JobModify_Rolling(t *testing.T) {
 func TestServiceSched_JobModify_Rolling_FullNode(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a node and clear the reserved resources
 	node := mock.Node()
@@ -2880,7 +2882,7 @@ func TestServiceSched_JobModify_Rolling_FullNode(t *testing.T) {
 func TestServiceSched_JobModify_Canaries(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -3004,7 +3006,7 @@ func TestServiceSched_JobModify_Canaries(t *testing.T) {
 func TestServiceSched_JobModify_InPlace(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -3154,7 +3156,7 @@ func TestServiceSched_JobModify_InPlace(t *testing.T) {
 func TestServiceSched_JobModify_InPlace08(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create node
 	node := mock.Node()
@@ -3233,7 +3235,7 @@ func TestServiceSched_JobModify_InPlace08(t *testing.T) {
 func TestServiceSched_JobModify_DistinctProperty(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -3347,7 +3349,7 @@ func TestServiceSched_JobModify_DistinctProperty(t *testing.T) {
 func TestServiceSched_JobModify_NodeReschedulePenalty(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -3475,7 +3477,7 @@ func TestServiceSched_JobModify_NodeReschedulePenalty(t *testing.T) {
 func TestServiceSched_JobDeregister_Purged(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Generate a fake job with allocations
 	job := mock.Job()
@@ -3544,7 +3546,7 @@ func TestServiceSched_JobDeregister_Purged(t *testing.T) {
 func TestServiceSched_JobDeregister_Stopped(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Generate a fake job with allocations
 	job := mock.Job()
@@ -3671,7 +3673,7 @@ func TestServiceSched_NodeDown(t *testing.T) {
 
 	for i, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
-			h := NewHarness(t)
+			h := tests.NewHarness(t)
 
 			// Register a node
 			node := mock.Node()
@@ -3789,7 +3791,7 @@ func TestServiceSched_StopOnClientAfter(t *testing.T) {
 
 	for i, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
-			h := NewHarness(t)
+			h := tests.NewHarness(t)
 
 			// Node, which is down
 			node := mock.Node()
@@ -3906,7 +3908,7 @@ func TestServiceSched_StopOnClientAfter(t *testing.T) {
 func TestServiceSched_NodeUpdate(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a node
 	node := mock.Node()
@@ -3962,7 +3964,7 @@ func TestServiceSched_NodeUpdate(t *testing.T) {
 func TestServiceSched_NodeDrain(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a draining node
 	node := mock.DrainNode()
@@ -4045,7 +4047,7 @@ func TestServiceSched_NodeDrain(t *testing.T) {
 func TestServiceSched_NodeDrain_Down(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a draining node
 	node := mock.DrainNode()
@@ -4158,7 +4160,7 @@ func TestServiceSched_NodeDrain_Down(t *testing.T) {
 
 func TestServiceSched_NodeDrain_Canaries(t *testing.T) {
 	ci.Parallel(t)
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	n1 := mock.Node()
 	n2 := mock.DrainNode()
@@ -4270,7 +4272,7 @@ func TestServiceSched_NodeDrain_Canaries(t *testing.T) {
 func TestServiceSched_NodeDrain_Queued_Allocations(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a draining node
 	node := mock.Node()
@@ -4323,8 +4325,8 @@ func TestServiceSched_NodeDrain_Queued_Allocations(t *testing.T) {
 func TestServiceSched_RetryLimit(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
-	h.Planner = &RejectPlan{h}
+	h := tests.NewHarness(t)
+	h.Planner = &tests.RejectPlan{h}
 
 	// Create some nodes
 	for i := 0; i < 10; i++ {
@@ -4375,7 +4377,7 @@ func TestServiceSched_RetryLimit(t *testing.T) {
 func TestServiceSched_Reschedule_OnceNow(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -4488,7 +4490,7 @@ func TestServiceSched_Reschedule_OnceNow(t *testing.T) {
 func TestServiceSched_Reschedule_Later(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 	// Create some nodes
 	var nodes []*structs.Node
 	for i := 0; i < 10; i++ {
@@ -4577,7 +4579,7 @@ func TestServiceSched_Reschedule_Later(t *testing.T) {
 func TestServiceSched_Reschedule_MultipleNow(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -4718,7 +4720,7 @@ func TestServiceSched_Reschedule_MultipleNow(t *testing.T) {
 func TestServiceSched_BlockedReschedule(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 	node := mock.Node()
 	must.NoError(t, h.State.UpsertNode(structs.MsgTypeTestSetup, h.NextIndex(), node))
 
@@ -4915,7 +4917,7 @@ func TestServiceSched_BlockedReschedule(t *testing.T) {
 func TestServiceSched_Reschedule_PruneEvents(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -5049,7 +5051,7 @@ func TestDeployment_FailedAllocs_Reschedule(t *testing.T) {
 
 	for _, failedDeployment := range []bool{false, true} {
 		t.Run(fmt.Sprintf("Failed Deployment: %v", failedDeployment), func(t *testing.T) {
-			h := NewHarness(t)
+			h := tests.NewHarness(t)
 			// Create some nodes
 			var nodes []*structs.Node
 			for i := 0; i < 10; i++ {
@@ -5134,7 +5136,7 @@ func TestDeployment_FailedAllocs_Reschedule(t *testing.T) {
 func TestBatchSched_Run_CompleteAlloc(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a node
 	node := mock.Node()
@@ -5193,7 +5195,7 @@ func TestBatchSched_Run_CompleteAlloc(t *testing.T) {
 func TestBatchSched_Run_FailedAlloc(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a node
 	node := mock.Node()
@@ -5265,7 +5267,7 @@ func TestBatchSched_Run_FailedAlloc(t *testing.T) {
 func TestBatchSched_Run_LostAlloc(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a node
 	node := mock.Node()
@@ -5354,7 +5356,7 @@ func TestBatchSched_Run_LostAlloc(t *testing.T) {
 func TestBatchSched_Run_FailedAllocQueuedAllocations(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	node := mock.DrainNode()
 	must.NoError(t, h.State.UpsertNode(structs.MsgTypeTestSetup, h.NextIndex(), node))
@@ -5408,7 +5410,7 @@ func TestBatchSched_Run_FailedAllocQueuedAllocations(t *testing.T) {
 func TestBatchSched_ReRun_SuccessfullyFinishedAlloc(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create two nodes, one that is drained and has a successfully finished
 	// alloc and a fresh undrained one
@@ -5483,7 +5485,7 @@ func TestBatchSched_ReRun_SuccessfullyFinishedAlloc(t *testing.T) {
 func TestBatchSched_JobModify_InPlace_Terminal(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -5537,7 +5539,7 @@ func TestBatchSched_JobModify_InPlace_Terminal(t *testing.T) {
 func TestBatchSched_JobModify_Destructive_Terminal(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	var nodes []*structs.Node
@@ -5623,7 +5625,7 @@ func TestBatchSched_JobModify_Destructive_Terminal(t *testing.T) {
 func TestBatchSched_NodeDrain_Running_OldJob(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create two nodes, one that is drained and has a successfully finished
 	// alloc and a fresh undrained one
@@ -5696,7 +5698,7 @@ func TestBatchSched_NodeDrain_Running_OldJob(t *testing.T) {
 func TestBatchSched_NodeDrain_Complete(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create two nodes, one that is drained and has a successfully finished
 	// alloc and a fresh undrained one
@@ -5762,7 +5764,7 @@ func TestBatchSched_NodeDrain_Complete(t *testing.T) {
 func TestBatchSched_ScaleDown_SameName(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a node
 	node := mock.Node()
@@ -5916,9 +5918,9 @@ func TestGenericSched_AllocFit_Lifecycle(t *testing.T) {
 	}
 	for _, testCase := range testCases {
 		t.Run(testCase.Name, func(t *testing.T) {
-			h := NewHarness(t)
+			h := tests.NewHarness(t)
 
-			legacyCpuResources, processorResources := cpuResources(testCase.NodeCpu)
+			legacyCpuResources, processorResources := tests.CpuResources(testCase.NodeCpu)
 			node := mock.Node()
 			node.NodeResources.Processors = processorResources
 			node.NodeResources.Cpu = legacyCpuResources
@@ -5967,7 +5969,7 @@ func TestGenericSched_AllocFit_Lifecycle(t *testing.T) {
 func TestGenericSched_AllocFit_MemoryOversubscription(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 	node := mock.Node()
 	node.NodeResources.Cpu.CpuShares = 10000
 	node.NodeResources.Memory.MemoryMB = 1224
@@ -6014,7 +6016,7 @@ func TestGenericSched_AllocFit_MemoryOversubscription(t *testing.T) {
 func TestGenericSched_ChainedAlloc(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	for i := 0; i < 10; i++ {
@@ -6050,7 +6052,7 @@ func TestGenericSched_ChainedAlloc(t *testing.T) {
 	sort.Strings(allocIDs)
 
 	// Create a new harness to invoke the scheduler again
-	h1 := NewHarnessWithState(t, h.State)
+	h1 := tests.NewHarnessWithState(t, h.State)
 	job1 := mock.Job()
 	job1.ID = job.ID
 	job1.TaskGroups[0].Tasks[0].Env["foo"] = "bar"
@@ -6105,7 +6107,7 @@ func TestGenericSched_ChainedAlloc(t *testing.T) {
 func TestServiceSched_NodeDrain_Sticky(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a draining node
 	node := mock.DrainNode()
@@ -6160,7 +6162,7 @@ func TestServiceSched_NodeDrain_Sticky(t *testing.T) {
 func TestServiceSched_CancelDeployment_Stopped(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Generate a fake job
 	job := mock.Job()
@@ -6236,7 +6238,7 @@ func TestServiceSched_CancelDeployment_Stopped(t *testing.T) {
 func TestServiceSched_CancelDeployment_NewerJob(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Generate a fake job
 	job := mock.Job()
@@ -6534,7 +6536,7 @@ func Test_updateRescheduleTracker(t *testing.T) {
 		t.Run(tc.desc, func(t *testing.T) {
 			prevAlloc.RescheduleTracker = &structs.RescheduleTracker{Events: tc.prevAllocEvents}
 			prevAlloc.Job.LookupTaskGroup(prevAlloc.TaskGroup).ReschedulePolicy = tc.reschedPolicy
-			updateRescheduleTracker(alloc, prevAlloc, tc.reschedTime)
+			UpdateRescheduleTracker(alloc, prevAlloc, tc.reschedTime)
 			must.Eq(t, tc.expectedRescheduleEvents, alloc.RescheduleTracker.Events)
 		})
 	}
@@ -6544,9 +6546,9 @@ func Test_updateRescheduleTracker(t *testing.T) {
 func TestServiceSched_Preemption(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
-	legacyCpuResources, processorResources := cpuResources(1000)
+	legacyCpuResources, processorResources := tests.CpuResources(1000)
 
 	// Create a node
 	node := mock.Node()
@@ -6707,7 +6709,7 @@ func TestServiceSched_Preemption(t *testing.T) {
 func TestServiceSched_Migrate_NonCanary(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	node1 := mock.Node()
 	must.NoError(t, h.State.UpsertNode(structs.MsgTypeTestSetup, h.NextIndex(), node1))
@@ -6779,7 +6781,7 @@ func TestServiceSched_Migrate_NonCanary(t *testing.T) {
 func TestServiceSched_Migrate_CanaryStatus(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	node1 := mock.Node()
 	must.NoError(t, h.State.UpsertNode(structs.MsgTypeTestSetup, h.NextIndex(), node1))
@@ -6948,7 +6950,7 @@ func TestServiceSched_Migrate_CanaryStatus(t *testing.T) {
 func TestDowngradedJobForPlacement_PicksTheLatest(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// This test tests downgradedJobForPlacement directly to ease testing many different scenarios
 	// without invoking the full machinary of scheduling and updating deployment state tracking.
@@ -7066,7 +7068,7 @@ func TestDowngradedJobForPlacement_PicksTheLatest(t *testing.T) {
 func TestServiceSched_RunningWithNextAllocation(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	node1 := mock.Node()
 	must.NoError(t, h.State.UpsertNode(structs.MsgTypeTestSetup, h.NextIndex(), node1))
@@ -7140,7 +7142,7 @@ func TestServiceSched_RunningWithNextAllocation(t *testing.T) {
 func TestServiceSched_CSIVolumesPerAlloc(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes, each running the CSI plugin
 	for i := 0; i < 5; i++ {
@@ -7307,7 +7309,7 @@ func TestServiceSched_CSIVolumesPerAlloc(t *testing.T) {
 
 func TestServiceSched_CSITopology(t *testing.T) {
 	ci.Parallel(t)
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	zones := []string{"zone-0", "zone-1", "zone-2", "zone-3"}
 
@@ -7414,7 +7416,7 @@ func TestServiceSched_Client_Disconnect_Creates_Updates_and_Evals(t *testing.T) 
 	for _, version := range jobVersions {
 		t.Run(version.name, func(t *testing.T) {
 
-			h := NewHarness(t)
+			h := tests.NewHarness(t)
 			count := 1
 			maxClientDisconnect := 10 * time.Minute
 
@@ -7492,7 +7494,7 @@ func TestServiceSched_Client_Disconnect_Creates_Updates_and_Evals(t *testing.T) 
 func TestServiceSched_ReservedCores_InPlace(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a node
 	node := mock.Node()
@@ -7585,7 +7587,7 @@ func TestServiceSched_ReservedCores_InPlace(t *testing.T) {
 	h.AssertEvalStatus(t, structs.EvalStatusComplete)
 }
 
-func initNodeAndAllocs(t *testing.T, h *Harness, job *structs.Job,
+func initNodeAndAllocs(t *testing.T, h *tests.Harness, job *structs.Job,
 	nodeStatus, clientStatus string) (*structs.Node, *structs.Job, []*structs.Allocation) {
 	// Node, which is ready
 	node := mock.Node()
