+++ scheduler/scheduler_sysbatch_test.go
@@ -15,6 +15,7 @@ import (
 	"github.com/hashicorp/nomad/helper/uuid"
 	"github.com/hashicorp/nomad/nomad/mock"
 	"github.com/hashicorp/nomad/nomad/structs"
+	"github.com/hashicorp/nomad/scheduler/tests"
 	"github.com/kr/pretty"
 	"github.com/shoenig/test/must"
 )
@@ -22,7 +23,7 @@ import (
 func TestSysBatch_JobRegister(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	_ = createNodes(t, h, 10)
@@ -95,7 +96,7 @@ func TestSysBatch_JobRegister(t *testing.T) {
 func TestSysBatch_JobRegister_AddNode_Running(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	nodes := createNodes(t, h, 10)
@@ -172,7 +173,7 @@ func TestSysBatch_JobRegister_AddNode_Running(t *testing.T) {
 func TestSysBatch_JobRegister_AddNode_Dead(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	nodes := createNodes(t, h, 10)
@@ -250,7 +251,7 @@ func TestSysBatch_JobRegister_AddNode_Dead(t *testing.T) {
 func TestSysBatch_JobModify(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	nodes := createNodes(t, h, 10)
@@ -340,7 +341,7 @@ func TestSysBatch_JobModify(t *testing.T) {
 func TestSysBatch_JobModify_InPlace(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	nodes := createNodes(t, h, 10)
@@ -414,7 +415,7 @@ func TestSysBatch_JobModify_InPlace(t *testing.T) {
 func TestSysBatch_JobDeregister_Purged(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	nodes := createNodes(t, h, 10)
@@ -475,7 +476,7 @@ func TestSysBatch_JobDeregister_Purged(t *testing.T) {
 func TestSysBatch_JobDeregister_Stopped(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	nodes := createNodes(t, h, 10)
@@ -538,7 +539,7 @@ func TestSysBatch_JobDeregister_Stopped(t *testing.T) {
 func TestSysBatch_NodeDown(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a down node
 	node := mock.Node()
@@ -599,7 +600,7 @@ func TestSysBatch_NodeDown(t *testing.T) {
 func TestSysBatch_NodeDrain_Down(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a draining node
 	node := mock.DrainNode()
@@ -653,7 +654,7 @@ func TestSysBatch_NodeDrain_Down(t *testing.T) {
 func TestSysBatch_NodeDrain(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a draining node
 	node := mock.DrainNode()
@@ -710,7 +711,7 @@ func TestSysBatch_NodeDrain(t *testing.T) {
 func TestSysBatch_NodeUpdate(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a node
 	node := mock.Node()
@@ -754,8 +755,8 @@ func TestSysBatch_NodeUpdate(t *testing.T) {
 func TestSysBatch_RetryLimit(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
-	h.Planner = &RejectPlan{h}
+	h := tests.NewHarness(t)
+	h.Planner = &tests.RejectPlan{h}
 
 	// Create some nodes
 	_ = createNodes(t, h, 10)
@@ -800,7 +801,7 @@ func TestSysBatch_RetryLimit(t *testing.T) {
 func TestSysBatch_Queued_With_Constraints(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	nodes := createNodes(t, h, 3)
 
@@ -847,7 +848,7 @@ func TestSysBatch_Queued_With_Constraints(t *testing.T) {
 func TestSysBatch_Queued_With_Constraints_PartialMatch(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// linux machines
 	linux := createNodes(t, h, 3)
@@ -897,7 +898,7 @@ func TestSysBatch_Queued_With_Constraints_PartialMatch(t *testing.T) {
 func TestSysBatch_JobConstraint_AddNode(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create two nodes
 	var node *structs.Node
@@ -1037,7 +1038,7 @@ func TestSysBatch_JobConstraint_AddNode(t *testing.T) {
 
 func TestSysBatch_JobConstraint_AllFiltered(t *testing.T) {
 	ci.Parallel(t)
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create two nodes, one with a custom class
 	node := mock.Node()
@@ -1083,7 +1084,7 @@ func TestSysBatch_JobConstraint_AllFiltered(t *testing.T) {
 
 func TestSysBatch_JobConstraint_RunMultiple(t *testing.T) {
 	ci.Parallel(t)
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create two nodes, one with a custom class
 	fooNode := mock.Node()
@@ -1153,7 +1154,7 @@ func TestSysBatch_JobConstraint_RunMultiple(t *testing.T) {
 func TestSysBatch_ExistingAllocNoNodes(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	var node *structs.Node
 	// Create a node
@@ -1234,7 +1235,7 @@ func TestSysBatch_ExistingAllocNoNodes(t *testing.T) {
 func TestSysBatch_ConstraintErrors(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	var node *structs.Node
 	// Register some nodes
@@ -1309,7 +1310,7 @@ func TestSysBatch_ConstraintErrors(t *testing.T) {
 func TestSysBatch_ChainedAlloc(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	_ = createNodes(t, h, 10)
@@ -1342,7 +1343,7 @@ func TestSysBatch_ChainedAlloc(t *testing.T) {
 	sort.Strings(allocIDs)
 
 	// Create a new harness to invoke the scheduler again
-	h1 := NewHarnessWithState(t, h.State)
+	h1 := tests.NewHarnessWithState(t, h.State)
 	job1 := mock.SystemBatchJob()
 	job1.ID = job.ID
 	job1.TaskGroups[0].Tasks[0].Env = make(map[string]string)
@@ -1398,7 +1399,7 @@ func TestSysBatch_ChainedAlloc(t *testing.T) {
 func TestSysBatch_PlanWithDrainedNode(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register two nodes with two different classes
 	node := mock.DrainNode()
@@ -1480,7 +1481,7 @@ func TestSysBatch_PlanWithDrainedNode(t *testing.T) {
 func TestSysBatch_QueuedAllocsMultTG(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register two nodes with two different classes
 	node := mock.Node()
@@ -1538,9 +1539,9 @@ func TestSysBatch_QueuedAllocsMultTG(t *testing.T) {
 func TestSysBatch_Preemption(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
-	legacyCpuResources, processorResources := cpuResources(3072)
+	legacyCpuResources, processorResources := tests.CpuResources(3072)
 
 	// Create nodes
 	nodes := make([]*structs.Node, 0)
@@ -1838,7 +1839,7 @@ func TestSysBatch_canHandle(t *testing.T) {
 		must.True(t, s.canHandle(structs.EvalTriggerPeriodicJob))
 	})
 }
-func createNodes(t *testing.T, h *Harness, n int) []*structs.Node {
+func createNodes(t *testing.T, h *tests.Harness, n int) []*structs.Node {
 	nodes := make([]*structs.Node, n)
 	for i := 0; i < n; i++ {
 		node := mock.Node()
