+++ scheduler/scheduler_system_test.go
@@ -16,14 +16,16 @@ import (
 	"github.com/hashicorp/nomad/helper/uuid"
 	"github.com/hashicorp/nomad/nomad/mock"
 	"github.com/hashicorp/nomad/nomad/structs"
+	"github.com/hashicorp/nomad/scheduler/feasible"
 	"github.com/hashicorp/nomad/scheduler/reconcile"
+	"github.com/hashicorp/nomad/scheduler/tests"
 	"github.com/shoenig/test/must"
 )
 
 func TestSystemSched_JobRegister(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	_ = createNodes(t, h, 10)
@@ -96,7 +98,7 @@ func TestSystemSched_JobRegister(t *testing.T) {
 func TestSystemSched_JobRegister_StickyAllocs(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	_ = createNodes(t, h, 10)
@@ -147,7 +149,7 @@ func TestSystemSched_JobRegister_StickyAllocs(t *testing.T) {
 		Status:      structs.EvalStatusPending,
 	}
 	must.NoError(t, h.State.UpsertEvals(structs.MsgTypeTestSetup, h.NextIndex(), []*structs.Evaluation{eval}))
-	h1 := NewHarnessWithState(t, h.State)
+	h1 := tests.NewHarnessWithState(t, h.State)
 	if err := h1.Process(NewSystemScheduler, eval); err != nil {
 		t.Fatalf("err: %v", err)
 	}
@@ -171,7 +173,7 @@ func TestSystemSched_JobRegister_StickyAllocs(t *testing.T) {
 func TestSystemSched_JobRegister_EphemeralDiskConstraint(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a node
 	node := mock.Node()
@@ -215,7 +217,7 @@ func TestSystemSched_JobRegister_EphemeralDiskConstraint(t *testing.T) {
 	}
 
 	// Create a new harness to test the scheduling result for the second job
-	h1 := NewHarnessWithState(t, h.State)
+	h1 := tests.NewHarnessWithState(t, h.State)
 	// Create a mock evaluation to register the job
 	eval1 := &structs.Evaluation{
 		Namespace:   structs.DefaultNamespace,
@@ -242,7 +244,7 @@ func TestSystemSched_JobRegister_EphemeralDiskConstraint(t *testing.T) {
 func TestSystemSched_ExhaustResources(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create a node
 	node := mock.Node()
@@ -320,7 +322,7 @@ func TestSystemSched_ExhaustResources(t *testing.T) {
 func TestSystemSched_JobRegister_Annotate(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	for i := 0; i < 10; i++ {
@@ -419,7 +421,7 @@ func TestSystemSched_JobRegister_Annotate(t *testing.T) {
 func TestSystemSched_JobRegister_AddNode(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	nodes := createNodes(t, h, 10)
@@ -499,7 +501,7 @@ func TestSystemSched_JobRegister_AddNode(t *testing.T) {
 func TestSystemSched_JobRegister_AllocFail(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create NO nodes
 	// Create a job
@@ -533,7 +535,7 @@ func TestSystemSched_JobRegister_AllocFail(t *testing.T) {
 func TestSystemSched_JobModify(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	nodes := createNodes(t, h, 10)
@@ -622,7 +624,7 @@ func TestSystemSched_JobModify(t *testing.T) {
 func TestSystemSched_JobModify_Rolling(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	nodes := createNodes(t, h, 10)
@@ -722,7 +724,7 @@ func TestSystemSched_JobModify_Rolling(t *testing.T) {
 func TestSystemSched_JobModify_InPlace(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	nodes := createNodes(t, h, 10)
@@ -804,7 +806,7 @@ func TestSystemSched_JobModify_InPlace(t *testing.T) {
 func TestSystemSched_JobModify_RemoveDC(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	node1 := mock.Node()
@@ -889,7 +891,7 @@ func TestSystemSched_JobModify_RemoveDC(t *testing.T) {
 func TestSystemSched_JobDeregister_Purged(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	nodes := createNodes(t, h, 10)
@@ -950,7 +952,7 @@ func TestSystemSched_JobDeregister_Purged(t *testing.T) {
 func TestSystemSched_JobDeregister_Stopped(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	nodes := createNodes(t, h, 10)
@@ -1013,7 +1015,7 @@ func TestSystemSched_JobDeregister_Stopped(t *testing.T) {
 func TestSystemSched_NodeDown(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a down node
 	node := mock.Node()
@@ -1074,7 +1076,7 @@ func TestSystemSched_NodeDown(t *testing.T) {
 func TestSystemSched_NodeDrain_Down(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a draining node
 	node := mock.DrainNode()
@@ -1128,7 +1130,7 @@ func TestSystemSched_NodeDrain_Down(t *testing.T) {
 func TestSystemSched_NodeDrain(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a draining node
 	node := mock.DrainNode()
@@ -1185,7 +1187,7 @@ func TestSystemSched_NodeDrain(t *testing.T) {
 func TestSystemSched_NodeUpdate(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a node
 	node := mock.Node()
@@ -1229,8 +1231,8 @@ func TestSystemSched_NodeUpdate(t *testing.T) {
 func TestSystemSched_RetryLimit(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
-	h.Planner = &RejectPlan{h}
+	h := tests.NewHarness(t)
+	h.Planner = &tests.RejectPlan{h}
 
 	// Create some nodes
 	_ = createNodes(t, h, 10)
@@ -1275,7 +1277,7 @@ func TestSystemSched_RetryLimit(t *testing.T) {
 func TestSystemSched_Queued_With_Constraints(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register a node
 	node := mock.Node()
@@ -1316,7 +1318,7 @@ func TestSystemSched_Queued_With_Constraints(t *testing.T) {
 func TestSystemSched_JobConstraint_AddNode(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create two nodes
 	var node *structs.Node
@@ -1458,7 +1460,7 @@ func TestSystemSched_JobConstraint_AddNode(t *testing.T) {
 
 func TestSystemSched_JobConstraint_AllFiltered(t *testing.T) {
 	ci.Parallel(t)
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create two nodes, one with a custom class
 	node := mock.Node()
@@ -1505,7 +1507,7 @@ func TestSystemSched_JobConstraint_AllFiltered(t *testing.T) {
 // subsequent runs, and report the outcome appropriately
 func TestSystemSched_JobConstraint_RunMultipleTimes(t *testing.T) {
 	ci.Parallel(t)
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create two nodes, one with a custom name
 	fooNode := mock.Node()
@@ -1545,14 +1547,14 @@ func TestSystemSched_JobConstraint_RunMultipleTimes(t *testing.T) {
 	testCases := []struct {
 		desc  string
 		job   func() *structs.Job
-		check func(*testing.T, *Harness)
+		check func(*testing.T, *tests.Harness)
 	}{
 		{
 			desc: "Rerunning the job with constraint shouldn't report failed allocations",
 			job: func() *structs.Job {
 				return job
 			},
-			check: func(t *testing.T, h *Harness) {
+			check: func(t *testing.T, h *tests.Harness) {
 				// Ensure a plan is not added, because no action should be taken
 				must.Len(t, 1, h.Plans)
 
@@ -1574,7 +1576,7 @@ func TestSystemSched_JobConstraint_RunMultipleTimes(t *testing.T) {
 				must.NoError(t, h.State.UpsertJob(structs.MsgTypeTestSetup, h.NextIndex(), nil, job))
 				return job
 			},
-			check: func(t *testing.T, h *Harness) {
+			check: func(t *testing.T, h *tests.Harness) {
 				// Ensure another plan is added for the updated alloc
 				must.Len(t, 2, h.Plans)
 
@@ -1601,7 +1603,7 @@ func TestSystemSched_JobConstraint_RunMultipleTimes(t *testing.T) {
 				must.NoError(t, h.State.UpsertJob(structs.MsgTypeTestSetup, h.NextIndex(), nil, job))
 				return job
 			},
-			check: func(t *testing.T, h *Harness) {
+			check: func(t *testing.T, h *tests.Harness) {
 				// Ensure another plan is added for the updated alloc
 				must.Len(t, 3, h.Plans)
 
@@ -1625,7 +1627,7 @@ func TestSystemSched_JobConstraint_RunMultipleTimes(t *testing.T) {
 				must.NoError(t, h.State.UpsertJob(structs.MsgTypeTestSetup, h.NextIndex(), nil, job))
 				return job
 			},
-			check: func(t *testing.T, h *Harness) {
+			check: func(t *testing.T, h *tests.Harness) {
 				// Ensure another Plan for the updated alloc
 				must.Len(t, 4, h.Plans)
 
@@ -1661,7 +1663,7 @@ func TestSystemSched_JobConstraint_RunMultipleTimes(t *testing.T) {
 func TestSystemSched_ExistingAllocNoNodes(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	var node *structs.Node
 	// Create a node
@@ -1742,7 +1744,7 @@ func TestSystemSched_ExistingAllocNoNodes(t *testing.T) {
 func TestSystemSched_ConstraintErrors(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	var node *structs.Node
 	// Register some nodes
@@ -1815,7 +1817,7 @@ func TestSystemSched_ConstraintErrors(t *testing.T) {
 func TestSystemSched_ChainedAlloc(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Create some nodes
 	_ = createNodes(t, h, 10)
@@ -1848,7 +1850,7 @@ func TestSystemSched_ChainedAlloc(t *testing.T) {
 	sort.Strings(allocIDs)
 
 	// Create a new harness to invoke the scheduler again
-	h1 := NewHarnessWithState(t, h.State)
+	h1 := tests.NewHarnessWithState(t, h.State)
 	job1 := mock.SystemJob()
 	job1.ID = job.ID
 	job1.TaskGroups[0].Tasks[0].Env = make(map[string]string)
@@ -1905,7 +1907,7 @@ func TestSystemSched_ChainedAlloc(t *testing.T) {
 func TestSystemSched_PlanWithDrainedNode(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register two nodes with two different classes
 	node := mock.DrainNode()
@@ -1987,7 +1989,7 @@ func TestSystemSched_PlanWithDrainedNode(t *testing.T) {
 func TestSystemSched_QueuedAllocsMultTG(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	// Register two nodes with two different classes
 	node := mock.Node()
@@ -2045,9 +2047,9 @@ func TestSystemSched_QueuedAllocsMultTG(t *testing.T) {
 func TestSystemSched_Preemption(t *testing.T) {
 	ci.Parallel(t)
 
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
-	legacyCpuResources, processorResources := cpuResources(3072)
+	legacyCpuResources, processorResources := tests.CpuResources(3072)
 
 	// Create nodes
 	nodes := make([]*structs.Node, 0)
@@ -3058,7 +3060,7 @@ func TestSystemSched_NodeDisconnected(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			h := NewHarness(t)
+			h := tests.NewHarness(t)
 
 			// Register a node
 			node := mock.Node()
@@ -3203,7 +3205,7 @@ func TestSystemSched_NodeDisconnected(t *testing.T) {
 
 func TestSystemSched_CSITopology(t *testing.T) {
 	ci.Parallel(t)
-	h := NewHarness(t)
+	h := tests.NewHarness(t)
 
 	zones := []string{"zone-0", "zone-1", "zone-2", "zone-3"}
 
@@ -3291,7 +3293,7 @@ func TestSystemSched_CSITopology(t *testing.T) {
 func TestEvictAndPlace_LimitLessThanAllocs(t *testing.T) {
 	ci.Parallel(t)
 
-	_, ctx := testContext(t)
+	_, ctx := feasible.MockContext(t)
 	allocs := []reconcile.AllocTuple{
 		{Alloc: &structs.Allocation{ID: uuid.Generate()}},
 		{Alloc: &structs.Allocation{ID: uuid.Generate()}},
@@ -3312,7 +3314,7 @@ func TestEvictAndPlace_LimitLessThanAllocs(t *testing.T) {
 func TestEvictAndPlace_LimitEqualToAllocs(t *testing.T) {
 	ci.Parallel(t)
 
-	_, ctx := testContext(t)
+	_, ctx := feasible.MockContext(t)
 	allocs := []reconcile.AllocTuple{
 		{Alloc: &structs.Allocation{ID: uuid.Generate()}},
 		{Alloc: &structs.Allocation{ID: uuid.Generate()}},
@@ -3332,7 +3334,7 @@ func TestEvictAndPlace_LimitEqualToAllocs(t *testing.T) {
 func TestEvictAndPlace_LimitGreaterThanAllocs(t *testing.T) {
 	ci.Parallel(t)
 
-	_, ctx := testContext(t)
+	_, ctx := feasible.MockContext(t)
 	allocs := []reconcile.AllocTuple{
 		{Alloc: &structs.Allocation{ID: uuid.Generate()}},
 		{Alloc: &structs.Allocation{ID: uuid.Generate()}},
